fastAPI/demo1/requirements.txt
aioredis==1.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.10.4
async-timeout==5.0.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.11
fastapi-cache2==0.2.2
greenlet==3.2.2
h11==0.14.0
hiredis==3.1.0
idna==3.10
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
redis==4.6.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
